refactor: split settings form into smaller components
