--- renderer/components/Settings/SettingsContext.js
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export default React.createContext({})
