+++ renderer/containers/Settings/SettingsForm.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 import SettingsForm from 'components/Settings/SettingsForm'
 import { saveConfigOverrides } from 'reducers/settings'
-import { showNotification } from 'reducers/notification'
+import { showNotification, showError } from 'reducers/notification'
 import { setLocale } from 'reducers/locale'
 import { configureAutoUpdater } from 'reducers/autoupdate'
 
@@ -9,6 +9,7 @@ const mapDispatchToProps = {
   configureAutoUpdater,
   saveConfigOverrides,
   showNotification,
+  showError,
   setLocale,
 }
